LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/symmetric-largeop-expected.txt

⫿
⫿

FAIL Placement of operators assert_equals: symmetric largeop is centered around the math axis expected 75 but got 0
